feat(mobile): share GPX via Web Share API with download fallback
